cluster-health-logger connects over unix socket
This is a local only component, so connect over a unix socket to allow
"require-secure-transport" to be enabled.  A local socket is somewhat
more efficient as well.

This change requires the cluster-health-logger job in pxc-release to
allow access to the MySQL Unix socket when this component is run under
bpm.

[#177349859](https://www.pivotaltracker.com/story/show/177349859)